requirements.txt
matplotlib==3.5.3
numpy==1.21.6
pandas==1.3.5
Pillow==9.4.0
plotly==5.13.1
pydeck==0.8.0
requests==2.28.2
seaborn==0.12.2
streamlit==1.20.0
